@@ -7,6 +7,8 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Senparc.CO2NET.RegisterServices;
+using Senparc.Weixin.RegisterServices;
 
 namespace MicroZero.Http.Gateway
 {
@@ -25,10 +27,13 @@ namespace MicroZero.Http.Gateway
         {
             IocHelper.SetServiceCollection(services);
 
+            services.AddMemoryCache();
+            services.AddSenparcGlobalServices(Configuration)//Senparc.CO2NET ȫ��ע��
+                    .AddSenparcWeixinServices(Configuration);//Senparc.Weixin ע��
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             app.UseDefaultFiles();
             app.UseStaticFiles();
@@ -36,7 +41,7 @@ namespace MicroZero.Http.Gateway
             app.UseFileServer();
 
             string ext = ConfigurationManager.AppSettings["WechatPath"];
-            LogRecorderX.SystemLog(ext);
+            LogRecorder.SystemLog(ext);
             RouteApp.Extends.Add(ext, () => new WechatRouter());
             RouteApp.Extends.Add(ext + "/", () => new WechatRouter());
             RouteApp.Initialize();
