@@ -294,7 +294,7 @@ namespace Agebull.MicroZero
             }
             catch (Exception e)
             {
-                LogRecorderX.Exception(e);
+                LogRecorder.Exception(e);
                 return false;
             }
         }
