@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.Runtime.Serialization;
 using Newtonsoft.Json;
 
@@ -7,9 +8,38 @@ namespace MicroZero.Http.Gateway
     /// <summary>
     ///     ��������
     /// </summary>
-    [DataContract]
     [JsonObject(MemberSerialization.OptIn)]
     public class CacheOption
+    {
+        /// <summary>
+        ///     ����API����
+        /// </summary>
+        [JsonProperty("api")]
+        public List<ApiCacheOption> Api { get; set; }
+
+
+        /// <summary>
+        ///     ˢ�´���API����
+        /// </summary>
+        [JsonProperty("trigger")]
+        public List<CacheFlushOption> Trigger { get; set; }
+
+        /// <summary>
+        ///     ��ʼ��
+        /// </summary>
+        public void Initialize()
+        {
+            Api?.ForEach(p => p.Initialize());
+        }
+    }
+
+
+    /// <summary>
+    ///     ��������
+    /// </summary>
+    [DataContract]
+    [JsonObject(MemberSerialization.OptIn)]
+    public class ApiCacheOption
     {
         /// <summary>
         ///     API����
@@ -17,6 +47,14 @@ namespace MicroZero.Http.Gateway
         [DataMember]
         [JsonProperty]
         public string Api { get; set; }
+
+
+        /// <summary>
+        ///     ����У�������ͷ
+        /// </summary>
+        [DataMember]
+        [JsonProperty]
+        public List<string> Keys { get; set; }
 
         /// <summary>
         ///     ����У�������ͷ
@@ -62,6 +100,11 @@ namespace MicroZero.Http.Gateway
                 FlushSecond = 300;
             else if (FlushSecond > 3600)
                 FlushSecond = 3600;
+            if (Keys != null && Keys.Count > 0)
+            {
+                Feature = CacheFeature.Keys;
+                return;
+            }
             if (Bear)
                 Feature |= CacheFeature.Bear;
             if (!OnlyName)
@@ -69,5 +112,36 @@ namespace MicroZero.Http.Gateway
             if (ByNetError)
                 Feature |= CacheFeature.NetError;
         }
+    }
+
+
+    /// <summary>
+    ///     ���津��ˢ������
+    /// </summary>
+    [DataContract]
+    [JsonObject(MemberSerialization.OptIn)]
+    public class CacheFlushOption
+    {
+        /// <summary>
+        ///     �������µ�API����
+        /// </summary>
+        [DataMember]
+        [JsonProperty]
+        public string TriggerApi { get; set; }
+
+        /// <summary>
+        ///     ��Ҫ�����API����
+        /// </summary>
+        [DataMember]
+        [JsonProperty]
+        public string CacheApi { get; set; }
+
+        /// <summary>
+        ///  �ֶ�ӳ��
+        /// </summary>
+        [DataMember]
+        [JsonProperty]
+        public Dictionary<string,string> Map { get; set; }
+
     }
 }
