@@ -3,7 +3,9 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading;
+using System.Threading.Tasks;
 using Agebull.Common.Logging;
+using Agebull.MicroZero;
 using Agebull.MicroZero.ZeroApis;
 
 namespace MicroZero.Http.Gateway
@@ -32,28 +34,57 @@ namespace MicroZero.Http.Gateway
         ///     ��黺��
         /// </summary>
         /// <returns>ȡ�����棬����ֱ�ӷ���</returns>
-        internal static bool LoadCache(RouteData data)
+        internal static async Task<bool> LoadCache(RouteData data)
         {
-            if (!CacheMap.TryGetValue(data.Uri.LocalPath, out data.CacheSetting))
+            var api = $"{data.ApiHost}/{data.ApiName}";
+            if (!CacheMap.TryGetValue(api, out data.CacheSetting))
             {
                 data.CacheKey = null;
                 return false;
             }
             var kb = new StringBuilder();
-            kb.Append(data.Uri.LocalPath);
-            if (data.CacheSetting.Feature.HasFlag(CacheFeature.Bear))
+            kb.Append(api);
+            kb.Append('?');
+            if (data.CacheSetting.Feature.HasFlag(CacheFeature.Keys))
             {
-                kb.Append(data.Token);
-            }
-            if (data.CacheSetting.Feature.HasFlag(CacheFeature.QueryString))
-            {
-                foreach (var kv in data.Arguments)
+                foreach (var key in data.CacheSetting.Keys)
                 {
-                    kb.Append($"&{kv.Key}={kv.Value}");
+                    if (data.Arguments.TryGetValue(key, out var value))
+                        kb.Append($"{key}={value}&");
                 }
-                if (!string.IsNullOrWhiteSpace(data.HttpContent))
+                try
                 {
-                    kb.Append(data.HttpContent);
+                    if (!string.IsNullOrWhiteSpace(data.HttpContent))
+                    {
+                        var dic = JsonHelper.DeserializeObject<Dictionary<string, string>>(data.HttpContent);
+
+                        foreach (var key in data.CacheSetting.Keys)
+                        {
+                            if (data.Arguments.TryGetValue(key, out var value))
+                                kb.Append($"{key}={value}&");
+                        }
+                    }
+                }
+                catch// (Exception)
+                {
+                }
+            }
+            else
+            {
+                if (data.CacheSetting.Feature.HasFlag(CacheFeature.Bear))
+                {
+                    kb.Append($"token={data.Token}&");
+                }
+                if (data.CacheSetting.Feature.HasFlag(CacheFeature.QueryString))
+                {
+                    foreach (var kv in data.Arguments)
+                    {
+                        kb.Append($"{kv.Key}={kv.Value}&");
+                    }
+                    if (!string.IsNullOrWhiteSpace(data.HttpContent))
+                    {
+                        kb.Append(data.HttpContent);
+                    }
                 }
             }
 
@@ -65,36 +96,28 @@ namespace MicroZero.Http.Gateway
                     IsLoading = 1,
                     Content = ApiResultIoc.NoReadyJson
                 });
-                LogRecorderX.MonitorTrace($"Cache Load {data.CacheKey}");
+                LogRecorder.MonitorTrace(() => $"Cache Load {data.CacheKey}");
                 return false;
             }
             if (cacheData.Success && (cacheData.UpdateTime > DateTime.Now || cacheData.IsLoading > 0))
             {
                 data.ResultMessage = cacheData.Content;
-                LogRecorderX.MonitorTrace($"Cache by {data.CacheKey}");
+                LogRecorder.MonitorTrace(() => $"Cache by {data.CacheKey}");
                 return true;
             }
             //һ�����룬�����ĵȴ����óɹ�
             if (Interlocked.Increment(ref cacheData.IsLoading) == 1)
             {
-                LogRecorderX.MonitorTrace($"Cache update {data.CacheKey}");
+                LogRecorder.MonitorTrace(() => $"Cache update {data.CacheKey}");
                 return false;
             }
-
+            Interlocked.Decrement(ref cacheData.IsLoading);
             //�ȴ����óɹ�
-            int cnt = 0;
-            while (++cnt < 10)
-            {
-                Thread.Sleep(50);
-                if (!Cache.TryGetValue(data.CacheKey, out var newData) || cacheData == newData)
-                    continue;
-                LogRecorderX.MonitorTrace($"Cache wait {data.CacheKey}");
-                data.ResultMessage = cacheData.Content;
-                return true;
-            }
-            LogRecorderX.MonitorTrace($"Cache Failed {data.CacheKey}");
-            data.CacheKey = null;
-            return false;
+            LogRecorder.MonitorTrace(() => $"Cache wait {data.CacheKey}");
+            var task = new TaskCompletionSource<string>();
+            cacheData.Waits.Add(task);
+            data.ResultMessage = await task.Task;
+            return true;
         }
 
         /// <summary>
@@ -103,20 +126,60 @@ namespace MicroZero.Http.Gateway
         /// <param name="data"></param>
         internal static void CacheResult(RouteData data)
         {
-            if (data.CacheSetting == null || data.CacheKey == null)
-                return;
-            if (!data.IsSucceed && !data.CacheSetting.Feature.HasFlag(CacheFeature.NetError)) 
-                return;
-            Cache[data.CacheKey]= new CacheData
+            if (data.CacheSetting != null && data.CacheKey != null)
             {
-                Content = data.ResultMessage,
-                Success = data.IsSucceed,
-                IsLoading = 0,
-                UpdateTime = DateTime.Now.AddSeconds(data.CacheSetting.FlushSecond)
-            };
-            LogRecorderX.MonitorTrace($"Cache succeed {data.CacheKey}");
-        }
+                if (Cache.TryGetValue(data.CacheKey, out var cd))
+                {
+                    var res = data.IsSucceed ? data.ResultMessage : cd.Content;
+                    foreach (var task in cd.Waits.ToArray())
+                    {
+                        task.TrySetResult(res);
+                    }
+                }
+                if (!data.IsSucceed && !data.CacheSetting.Feature.HasFlag(CacheFeature.NetError))
+                    return;
+                Cache[data.CacheKey] = new CacheData
+                {
+                    Content = data.ResultMessage,
+                    Success = data.IsSucceed,
+                    IsLoading = 0,
+                    UpdateTime = DateTime.Now.AddSeconds(data.CacheSetting.FlushSecond)
+                };
+                LogRecorder.MonitorTrace($"Cache succeed {data.CacheKey}");
+            }
+            if (data.IsSucceed && UpdateMap.TryGetValue($"{data.ApiHost}/{data.ApiName}", out var uc))
+            {
+                var kb = new StringBuilder();
+                kb.Append(uc.CacheApi);
+                kb.Append('?');
 
+                foreach (var map in uc.Map)
+                {
+                    if (data.Arguments.TryGetValue(map.Key, out var value))
+                        kb.Append($"{map.Value}={value}&");
+                }
+
+                //var jObject = (JObject)JsonConvert.DeserializeObject(data.ResultMessage);
+
+                //foreach (var map in uc.Map)
+                //{
+                //    if (jObject.TryGetValue(map.Key, out var value))
+                //        kb.Append($"{map.Value}={value}&");
+                //}
+                RemoveCache(kb.ToString());
+            }
+        }
+        public static void RemoveCache(string key)
+        {
+            if (Cache.TryRemove(key, out var data))
+            {
+                LogRecorder.MonitorTrace($"Cache remove {key}");
+                foreach (var task in data.Waits.ToArray())
+                {
+                    task.TrySetResult(data.Content);
+                }
+            }
+        }
         #endregion
 
         #region ����
@@ -124,7 +187,12 @@ namespace MicroZero.Http.Gateway
         /// <summary>
         ///     ·������
         /// </summary>
-        public static Dictionary<string, CacheOption> CacheMap { get; set; }
+        public static Dictionary<string, ApiCacheOption> CacheMap { get; set; }
+
+        /// <summary>
+        ///     ·������
+        /// </summary>
+        public static Dictionary<string, CacheFlushOption> UpdateMap { get; set; }
 
 
         /// <summary>
@@ -133,17 +201,27 @@ namespace MicroZero.Http.Gateway
         /// <returns></returns>
         public static void InitCache()
         {
-            CacheMap = new Dictionary<string, CacheOption>(StringComparer.OrdinalIgnoreCase);
-            if (RouteOption.Option.CacheSettings == null)
-                return;
-            foreach (var setting in RouteOption.Option.CacheSettings)
-            {
-                setting.Initialize();
-                if (!CacheMap.ContainsKey(setting.Api))
-                    CacheMap.Add(setting.Api, setting);
-                else
-                    CacheMap[setting.Api] = setting;
-            }
+            Cache.Clear();
+            CacheMap = new Dictionary<string, ApiCacheOption>(StringComparer.OrdinalIgnoreCase);
+            if (GatewayOption.Option.CacheSettings?.Api != null)
+                foreach (var setting in GatewayOption.Option.CacheSettings.Api)
+                {
+                    setting.Initialize();
+                    if (!CacheMap.ContainsKey(setting.Api))
+                        CacheMap.Add(setting.Api, setting);
+                    else
+                        CacheMap[setting.Api] = setting;
+                }
+
+            UpdateMap = new Dictionary<string, CacheFlushOption>(StringComparer.OrdinalIgnoreCase);
+            if (GatewayOption.Option.CacheSettings?.Trigger != null)
+                foreach (var setting in GatewayOption.Option.CacheSettings.Trigger)
+                {
+                    if (!UpdateMap.ContainsKey(setting.TriggerApi))
+                        UpdateMap.Add(setting.TriggerApi, setting);
+                    else
+                        UpdateMap[setting.TriggerApi] = setting;
+                }
         }
 
         #endregion
