@@ -13,6 +13,7 @@ namespace MicroZero.Http.Gateway
     /// </summary>
     public class GatewayStartup
     {
+
         /// <summary>
         /// 构造
         /// </summary>
@@ -34,23 +35,14 @@ namespace MicroZero.Http.Gateway
         public void ConfigureServices(IServiceCollection services)
         {
             IocHelper.SetServiceCollection(services);
+            ZeroApplication.WorkModel = ZeroWorkModel.Client;
+
             ZeroApplication.CheckOption();
-            RouteOption.CheckOption();
-
-            services.AddCors(options =>
-                options.AddPolicy("AllowSameDomain",
-                    builder => builder.WithOrigins("*").AllowAnyMethod().AllowAnyHeader().AllowAnyOrigin().AllowCredentials())
-            );
-
-            DoConfigureServices(services);
-        }
-
-        /// <summary>
-        /// 执行系统配置
-        /// </summary>
-        /// <param name="services"></param>
-        protected virtual void DoConfigureServices(IServiceCollection services)
-        {
+            GatewayOption.CheckOption();
+            //services.AddCors(options =>
+            //    options.AddPolicy("AllowSameDomain",
+            //        builder => builder.WithOrigins("*").AllowAnyMethod().AllowAnyHeader().AllowAnyOrigin().AllowCredentials())
+            //);
         }
 
         /// <summary>
@@ -65,19 +57,8 @@ namespace MicroZero.Http.Gateway
             app.UseCookiePolicy();
             app.UseFileServer();
 
-            DoConfigure(app, env);
-
             RouteApp.Initialize();
             app.Run(RouteApp.Call);
-        }
-
-        /// <summary>
-        /// 应用配置
-        /// </summary>
-        /// <param name="app"></param>
-        /// <param name="env"></param>
-        protected virtual void DoConfigure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
         }
     }
 
