@@ -1,9 +1,15 @@
 using System.Threading.Tasks;
 using Agebull.Common.Configuration;
+using Agebull.Common.Logging;
 using Agebull.MicroZero;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
-using ZeroMQ;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging.Configuration;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using System.IO;
+using System;
 
 namespace MicroZero.Http.Gateway
 {
@@ -12,8 +18,9 @@ namespace MicroZero.Http.Gateway
 
         public static async Task Main(string[] args)
         {
+            LogRecorder.LogPath = Path.Combine(Environment.CurrentDirectory, "logs", ConfigurationManager.Root["AppName"]);
             BuildWebHost(args).Run();
-            
+
             await ZeroApplication.Shutdown();
         }
 
@@ -21,7 +28,13 @@ namespace MicroZero.Http.Gateway
         {
             return WebHost.CreateDefaultBuilder(args)
                 .UseConfiguration(ConfigurationManager.Root)
-                
+                .ConfigureLogging((hostingContext, builder) =>
+                {
+                    builder.AddConfiguration(ConfigurationManager.Root.GetSection("Logging"));
+                    //builder.AddConsole();
+                    builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<ILoggerProvider, TextLoggerProvider>());
+                    LoggerProviderOptions.RegisterProviderOptions<TextLoggerOption, TextLoggerProvider>(builder.Services);
+                })
                 .UseKestrel(RouteApp.Options)//HTTPS�˿�����
                 .UseStartup<GatewayStartup>()
                 .Build();
